requirements.txt
click>=8.1.0
PyYAML>=6.0
watchdog>=4.0.0
